Add files via upload
ADDED: Weapons table which should include all the pistols and can just be commented out if they don't want those to work. -- JINK
MODIFIED: Spacing of the client.lua so that it should reflect accurately on github now.
ADDED: functions to check for weapons and moved disabling actions to a function.
BUGS: I added disabling firing of weapons to the function because even though you set the peds to unarmed they could still fire their weapons when talking on radio. This should fix that. --JINK